Split CyclingMetrics.jsx
Fixed an issue where I was accidentally routing to Hiking metrics components when clicking on the cycling category, whoops... lol

Another thing that is worth pointing out as I split the rest of these metric components is that the input and viewing on the graphs won't work since both components exist in separate files and the necessary state isn't being shared. Inputting and viewing won't fully work until our DB/API implementation is in place. No going back now.

Of course these two components will eventually live on separate pages anyways, so the app as we know it is going through some big changes :)
